allow pagePath to be set on a per-request basis instead of globally
Code that is separated out into modules require the use of multiple pagePaths, so having one global wasn't working.

In addition to that, I've added a context field that can give a sort of namespace to the compiled folder allowing for multiple files to be rendered when they share the same name. (e.g. users/index and groups/index)